user guard conditions for photo CRUD

+++ app/controllers/photos_controller.rb
@@ -1,4 +1,6 @@
 class PhotosController < ApplicationController
+  before_action :authenticate_user!, :except => [:index, :show]
+
   def index
     @photos = Photo.all.reverse
   end
@@ -22,14 +24,13 @@ class PhotosController < ApplicationController
 
   def update
     photo = Photo.find(params[:id])
-    photo.update(photo_params)
+    photo.update(photo_params) if photo.belong_to?(current_user)
     redirect_to photos_path
   end
 
   def destroy
     photo = Photo.find(params[:id])
-    photo.destroy
-    flash[:notice] = "Photo deleted!"
+    photo.destroy if photo.belong_to?(current_user)
     redirect_to photos_path
   end
